# static/openapi.yaml
openapi: 3.0.0
info:
  title: DupImg API
  version: 1.0.0
  description: 画像重複検出アプリケーションの API ドキュメント

paths:
  /api/data:
    get:
      summary: 進捗データを取得
      description: 現在の進捗状況を JSON 形式で返します。
      responses:
        '200':
          description: 進捗データを返します。
          content:
            application/json:
              schema:
                type: object
                properties:
                  current_time:
                    type: string
                    format: date-time
                    description: 現在時刻
                  elapsed_time:
                    type: string
                    description: 処理の経過時間
                  status:
                    type: string
                    description: 現在のステータス
                  message:
                    type: string
                    description: 現在のメッセージ
                  steps:
                    type: array
                    description: 処理ステップの進捗状況
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                          description: ステップ名
                        progress:
                          type: integer
                          description: ステップの進捗率（0～100）
                        status:
                          type: string
                          description: ステップのステータス（未開始、進行中、完了）
                  group_list:
                    type: array
                    description: 類似画像のグループリスト
                    items:
                      type: array
                      items:
                        type: object
                        properties:
                          paths:
                            type: array
                            items:
                              type: string
                            description: 画像のパス
                          size:
                            type: integer
                            description: 画像のサイズ（バイト単位）
                          date:
                            type: string
                            format: date-time
                            description: 画像の日付
                          dateType:
                            type: string
                            description: 日付の種類（exif または file）
                          hardlink_ability:
                            type: boolean
                            description: ハードリンク可能かどうか
                          device:
                            type: integer
                            description: デバイス ID
                          thumbnail:
                            type: string
                            format: byte
                            description: サムネイル画像の Base64 エンコード
        '500':
          description: サーバーエラー

  /api/list_directories:
    get:
      summary: サブディレクトリのリストを取得
      description: 指定されたディレクトリのサブディレクトリをリストします。
      parameters:
        - name: path
          in: query
          required: false
          description: 基準となるディレクトリのパス
          schema:
            type: string
      responses:
        '200':
          description: サブディレクトリのリストを返します。
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        '400':
          description: 無効なディレクトリパス

  /api/add_directory:
    post:
      summary: ディレクトリを追加
      description: 手入力されたディレクトリを追加します。
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                directory:
                  type: string
                  description: 追加するディレクトリのパス
      responses:
        '200':
          description: ディレクトリが正常に追加されました。
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    description: 処理結果のステータス
                  message:
                    type: string
                    description: 処理結果のメッセージ
        '400':
          description: 無効なディレクトリパス

  /api/image:
    get:
      summary: 画像を取得
      description: 指定されたパスの画像を返します。
      parameters:
        - name: path
          in: query
          required: true
          description: 画像ファイルの絶対パス
          schema:
            type: string
      responses:
        '200':
          description: 画像ファイル
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '404':
          description: ファイルが見つからない場合
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    description: エラーメッセージ